# api-docs.json
{"openapi":"3.0.1","info":{"title":"Library System","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/books/{id}":{"get":{"tags":["book-controller"],"operationId":"getBookById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["book-controller"],"operationId":"updateBookById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["book-controller"],"operationId":"deleteBookById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/users/":{"post":{"tags":["user-controller"],"operationId":"addOne","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/v1/categories":{"get":{"tags":["category-controller"],"operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Category"}}}}}}},"post":{"tags":["category-controller"],"operationId":"addOne_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Category"}}}}}}},"/api/v1/books":{"get":{"tags":["book-controller"],"operationId":"getAll_2","parameters":[{"name":"search","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookDTO"}}}}}}},"post":{"tags":["book-controller"],"operationId":"addOne_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/books/{id}/return":{"post":{"tags":["book-controller"],"operationId":"returnBook","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/books/{id}/borrow":{"post":{"tags":["book-controller"],"operationId":"borrowBook","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/authors":{"get":{"tags":["author-controller"],"operationId":"getAll_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["author-controller"],"operationId":"addOne_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/signin":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/users":{"get":{"tags":["user-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"/api/v1/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/users/{id}/loans":{"get":{"tags":["user-controller"],"operationId":"getLoansByUserId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserLoanDTO"}}}}}}}},"/api/v1/myaccount/loans":{"get":{"tags":["account-controller"],"operationId":"myLoans","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/categories/{id}":{"get":{"tags":["category-controller"],"operationId":"getCategoryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Category"}}}}}},"delete":{"tags":["category-controller"],"operationId":"deleteCategoryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/authors/{id}":{"get":{"tags":["author-controller"],"operationId":"getAuthorById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["author-controller"],"operationId":"deleteAuthorById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/userinfo":{"get":{"tags":["auth-controller"],"operationId":"userinfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"BookDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"isbn":{"type":"integer","format":"int64"},"title":{"type":"string"},"publisher":{"type":"string"},"description":{"type":"string"},"publishedDate":{"type":"string"},"bookCoverLink":{"type":"string"},"status":{"type":"string"},"category":{"type":"string"},"authors":{"type":"array","items":{"type":"string"}}}},"Author":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"books":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}},"Book":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"publisher":{"type":"string"},"publishedDate":{"type":"string","format":"date-time"},"description":{"type":"string"},"bookCoverLink":{"type":"string"},"status":{"type":"string","enum":["BORROWED","AVAILABLE"]},"category":{"$ref":"#/components/schemas/Category"},"authors":{"type":"array","items":{"$ref":"#/components/schemas/Author"}},"fields":{"$ref":"#/components/schemas/Book"},"isbn":{"type":"integer","format":"int64"}}},"Category":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"}}},"Loan":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"user":{"$ref":"#/components/schemas/User"},"book":{"$ref":"#/components/schemas/Book"},"borrowDate":{"type":"string","format":"date-time"},"returnByDate":{"type":"string","format":"date-time"},"returnedDate":{"type":"string","format":"date-time"}}},"User":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"password":{"type":"string","writeOnly":true},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN"]},"loans":{"type":"array","items":{"$ref":"#/components/schemas/Loan"}},"activeLoans":{"type":"array","items":{"$ref":"#/components/schemas/Loan"}}}},"UserDTO":{"type":"object","properties":{"username":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"loans":{"type":"array","items":{"$ref":"#/components/schemas/UserLoanDTO"}}}},"UserLoanDTO":{"type":"object","properties":{"borrowDate":{"type":"string","format":"date-time"},"returnByDate":{"type":"string","format":"date-time"},"book":{"$ref":"#/components/schemas/BookDTO"}}},"AuthRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}}